relocated ticket and session router to user components folder
